Add basic patches for FDCAN
FDCAN appears in the H7, G4, L5 and MP1 families. The peripheral varies between
the families, with the G4/L5 versions being somewhat simpler. I don't see any
variation within families, but haven't exhaustively checked for this.

On the H7, the FDCAN2 doesn't support TTCAN. However FDCAN2 is
`derivedFrom="FDCAN1"` in the original ST SVDs (and remains so). The RM isn't
very clear about which registers are missing from FDCAN2, but we could to remove
the `TT*` registers from FDCAN2 if this is important.

G4 devices have either 1 or 3 FDCANs.

Checked against:
H7 RM0399
G4 RM0440